[JAX Runner] Refactor jax runner to delegate persistent batch handling logic. [2/N]
